requirement.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==2.2.3
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
category-encoders==2.6.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
duckdb==1.1.3
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.1.0
Faker==19.13.0
fastjsonschema==2.20.0
filelock==3.16.1
flasgger==0.9.7.1
flask==2.2.5
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-restx==1.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kagglehub==0.3.4
kaleido==0.2.1
kiwisolver==1.4.7
lazy_loader==0.4
ldap3==2.9.1
lightgbm==4.5.0
llvmlite==0.43.0
locust==2.32.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
meraki==1.52.0
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
msal==1.31.1
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.1.0
narwhals==1.13.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
npm==0.1.1
numba==0.60.0
openai==1.54.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
optional-django==0.1.0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandasai==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
polars==1.13.1
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pycaret==3.3.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-magic==0.4.27
Pygments==2.18.0
PyJWT==1.7.1
PyMuPDF==1.25.1
PyMySQL==1.1.1
pyod==2.0.2
pyodbc==5.2.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyperclip==1.9.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.17
python-pptx==1.0.2
pytz==2024.2
python-dotenv==1.0.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
schemdraw==0.15
scikit-base==0.7.8
scikit-image==0.24.0
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.0.6
six==1.16.0
sktime==0.26.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlglot==25.29.0
sqlglotrs==0.2.13
stack-data==0.6.3
starlette==0.41.2
statsmodels==0.14.4
sympy==1.13.1
tbats==1.1.3
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tinycss2==1.4.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
tsdownsample==0.1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
pymssql==2.3.2
widgetsnbextension==4.0.13
xgboost==2.1.3
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.17.1
six==1.17.0
yellowbrick==1.5
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
